pyproject.toml
[tool.poetry]
name = "flatlander"
version = "0.1.0"
description = "An Ai Agent Space (named in honor of Beowulf Schaffer)"
authors = ["Your Name <you@example.com>"]
license = "GPL 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-community = "^0.2.7"
psycopg = "^3.2.1"
ollama = "^0.2.1"
litellm = "^1.41.21"
chainlit = "^1.1.306"
langchain-openai = "^0.1.17"
streamlit = "^1.36.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
